requirements.txt
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
factory-boy==2.10.0
